Calculate Burt constraint for the egos of ego-centered networks

This calculates Burt's network constraint for all egos in an egor object. It iterates over each network and applies igraph::constraint. A weight variable can be specified.

ego_constraint(object, weights = NULL, ego.alter.weights = weights)

Arguments

object
An egor object.

weights
Character, naming the alter-alter tie weight variable.

ego.alter.weights
Character, naming the ego-alter weight tie weight variable. This defaults to the same value as weights, only specify if the name of the ego.alter.weights is different from weights.

Value

Numeric vector with a constraint value for each ego.

Details

The calculation of weighted network constraint only works, if the alter-alter tie weights are complemented by a alter level variable specifying the same weight for the ego-alter ties.

References

Burt, R. (2004). Structural holes and good ideas. American Journal of Sociology, (110), 349–399.

Examples

data(egor32)
ego_constraint(egor32)

#> # A tibble: 32 x 2
#>    .egoID constraint
#>    <fct>       <dbl>
#>  1 1           0.397
#>  2 2           0.191
#>  3 3           0.226
#>  4 4           0.212
#>  5 5           0.329
#>  6 6           0.216
#>  7 7           0.311
#>  8 8           0.209
#>  9 9           0.346
#> 10 10          0.291
#> # … with 22 more rows