Display names and values of global egor options.

Display names and values of global egor options.

Usage

egor_options()

Details

Currently egor makes use of the following global options. Use options() to change values.

egor.print.rows.active.level

Numeric. Amount of rows to display when printing the active level of an egor object.

egor.print.rows.inactive.level

Numeric. Amount of rows to display when printing the inactive levels of an egor object.

egor.print.switch.active.level.to.top

Logical. When printing an egor object, should the active data-level always be printed first?

egor.return.results.with.design

Logical. egor functions that return ego-level results (e.g. one value per ego) return a svy_tbl object containing the ego_design(), when this is set to TRUE.