General helper functions

Helper functions for ego centered network analysis

Usage

as_nested_egor(x)

alters_by_ego(x)

# S3 method for class 'egor'
alters_by_ego(x)

# S3 method for class 'nested_egor'
alters_by_ego(x)

aaties_by_ego(x)

# S3 method for class 'egor'
aaties_by_ego(x)

# S3 method for class 'nested_egor'
aaties_by_ego(x)

dyad.poss(max.alters, directed = FALSE)

sanitize.wide.edges(max.alters)

create_edge_names_wide(x)

dyads_possible_between_groups(x, y, geometric = TRUE)

din_page_dist(x)

Arguments

x

Numeric.

max.alters

A numeric giving the maximum number of alters.

directed

A logical value indicating directedness of alter-alter data.

y

Numeric.

geometric

Logical. Calculate possible dyads for geometric mean?

Functions

as_nested_egor(): Converts an egor object to a "legacy" egor object with nested .alts and .aaties columns.

alters_by_ego(): Splits the alter table into a list of tables (possibly 0-row) of alters associated with each ego, in the same order as the ego table.

aaties_by_ego(): Splits the alter–alter ties table into a list of tables (possibly 0-row) of alter–alter associated with each ego, in the same order as the ego table.

dyad.poss(): Returns the count of possible edges in an un-directed or directed, ego-centered network, based on the number of alters.

sanitize.wide.edges(): Generates a data.frame marking possible dyads in a wide alter-alter relation data.frame. Row names corresponds to the network size. This is useful for sanitizing alter-alter relations in the wide format.

create_edge_names_wide(): Creates a vector of names for variables containing data on alter-alter relations/ dyads in ego-centered networks.

dyads_possible_between_groups(): Calculates the possible edges between members of different groups in an ego-centered network.

din_page_dist(): Calculates the optimal distribution of a number of equally sized objects on a DIN-Norm DIN 476 (i.e. DIN A4) page in landscape view.