Plotting egor objects

egor Objects can be plotted as egographs or egograms. By default networks of the four first egos are plotted.

plot_egograms(
  x,
  ego_no = 1,
  x_dim = 1,
  y_dim = 1,
  venn_var,
  pie_var,
  vertex_size_var = NULL,
  vertex_color_var = NULL,
  vertex_color_palette = "Heat Colors",
  vertex_color_legend_label = vertex_color_var,
  vertex_label_var = NULL,
  edge_width_var = NULL,
  edge_color_var = NULL,
  edge_color_palette = "Heat Colors",
  highlight_box_col_var = NULL,
  highlight_box_col_palette = "Heat Colors",
  res_disp_vars = NULL,
  vertex_zoom = 1,
  edge_zoom = 2,
  font_size = 1,
  venn_colors = NULL,
  show_venn_labels = TRUE,
  ...
)

plot_ego_graphs(
  x,
  ego_no = 1,
  x_dim = 1,
  y_dim = 1,
  vertex_size_var = NULL,
  vertex_color_var = NULL,
  vertex_color_palette = "Heat Colors",
  vertex_color_legend_label = vertex_color_var,
  vertex_label_var = NULL,
  edge_width_var = NULL,
  edge_color_var = NULL,
  edge_color_palette = "Heat Colors",
  highlight_box_col_var = NULL,
  highlight_box_col_palette = "Heat Colors",
  res_disp_vars = NULL,
  vertex_zoom = 1,
  edge_zoom = 3,
  font_size = 1,
  include_ego = FALSE,
  ...
)

plot_egor(
  x,
  ego_no = 1,
  x_dim = 2,
  y_dim = 2,
  ...,
  type = c("egograph", "egogram")
)

# S3 method for egor
plot(x, ...)

Arguments

x
An egor object.

ego_no
Ego row number.

x_dim
Number of ego networks to plot horizontally.

y_dim
Number of ego networks to plot vertically

venn_var
Name (character) of alter column.

pie_var
Name (character) of alter column.

vertex_size_var
Name (character) of alter column.

vertex_color_var
Name (character) of alter column.

vertex_color_palette
Name (character) of color palette.

vertex_color_legend_label
Character.

vertex_label_var
Name (character) of alter column.

edge_width_var
Name (character) of aatie column.

edge_color_var
Name (character) of aatie column.

edge_color_palette
Name (character) of color palette.

highlight_box_col_var
Name (character) of ego column.

highlight_box_col_palette
Name (character) of color palette.

res_disp_vars
Name (character) of ego column.

vertex_zoom
Numeric.

edge_zoom
Numeric.

font_size
Numeric.

venn_colors
Vector of colors.

show_venn_labels
Logical.

...
Additional arguments forwared to plot.igraph.

include_ego
Logical.

type
Character. Either "egograph" or "egogram".

Details

For type eqals "egograph" ego networks are plotted using

Functions

plot_egograms: Plots an ego-socio-gram.

plot_ego_graphs: Plots an ego graph.