Calculate the EI-Index for the alter-alter ties of an ego object

The EI-Index is the division of the intra-group edge density and the outer-group edge density. It is calculated for the whole network and for subgroups. The whole network EI is a metric indicating the tendency of a network to be clustered by the categories of a given factor variable. The EI value of a groups describes the tendency of a group to be connected or not connected to other groups. Additionally, the EI index can be employed as a measurement for egos tendency to homo-/heteorphily - use the comp_ei() command for that version of EI-Index.

EI(object, alt.attr)

Arguments

object
An egor object.

alt.attr
Character naming grouping variable.

References

Krackhardt, D., Stern, R.N., 1988. Informal networks and organizational crises: an experimental simulation. Social Psychology Quarterly 51 (2), 123-140.

Everett, M. G., & Borgatti, S. P. (2012). Categorical attribute based centrality: E-I and G-F centrality. Social Networks, 34(4), 562-569.

Examples

data("egor32")
EI(egor32, "sex")

#> EI-Index: sex

#> # A tibble: 32 x 3
#>      ei_sc       m       w
#>      <dbl>   <dbl>   <dbl>
#>  1  0.28    0.250   1     
#>  2  0.0537  0.0430  0.0886
#>  3 -0.110  -0.0849 -0.151 
#>  4  0.0198  0.125  -0.150 
#>  5  0.0191  0       0.176 
#>  6 -0.160  -0.101  -0.245 
#>  7  0.0931  0.0588  0.233 
#>  8  0.0820  0.111   0.0244
#>  9  0.158   0.2    -0.333 
#> 10 -0.0581  0.0588 -0.321 
#> # … with 22 more rows