Convert egor object to network or igraph objects

These functions convert an egor object into a list of network or igraph objects. By default ego itself is not included in the created objects, there is a parameter (include.egor) that allows for including ego.

as_igraph(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)

# S3 method for nested_egor
as_igraph(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)

as.igraph.egor(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)

Arguments

x
An egor object.

directed
Logical, indicating if alter-alter relations are directed.

include.ego
Logical. Should ego be included?

ego.attrs
Vector of names (character) or indices (numeric) of ego variables that should be carried over to the network/ igraph objects.

ego.alter.weights
Vector of names (character) or indices (numeric) of alter variables that should be carried over to the the network/ igraph objects, as edge attributes of the ego-alter relations.

Details

The names of the variables specified in ego.attr and ego.alter.attr need to be the same as the names of corresponding alter attributes, in order for those variables to be complete in the resulting network/ igraph object (see example).