Calculate third-party compositional measures on an egor object

comp_ply() applies a function, that uses an alter attribute to calculate a compositional measurement, on all networks in an egor object and returns a numeric vector.

comp_ply(object, alt.attr, .f, ..., ego.attr = NULL)

Arguments

object
An egor object.

alt.attr
A character naming the variable containing the alter-attribute.

.f
A function that returns a numeric.

...
Optional arguments to .f.

ego.attr
Optional character naming an ego attribute.

Value

A numeric vector.

Details

When an ego attribute is used the .f is called like this: .f(alt.attr, ego.attr, ...). .f must return a single numeric value.

Examples

df <- make_egor(10, 32)
comp_ply(df, "age.years", sd, na.rm = TRUE)

#> # A tibble: 10 x 2
#>    .egoID result
#>    <chr>   <dbl>
#>  1 1        34.0
#>  2 2        32.3
#>  3 3        28.4
#>  4 4        30.9
#>  5 5        31.7
#>  6 6        33.3
#>  7 7        30.6
#>  8 8        30.9
#>  9 9        29.5
#> 10 10       31.1