Set and query the ego sampling design

Extract, set, remove, or update the survey design associated with an ego-centered dataset.

ego_design(x, ...)

# S3 method for egor
ego_design(x, ...)

# S3 method for nested_egor
ego_design(x, ...)

ego_design(x, ...) <- value

# S3 method for egor
ego_design(x, ...) <- value

# S3 method for nested_egor
ego_design(x, ...) <- value

has_ego_design(x)

# S3 method for egor
has_ego_design(x)

# S3 method for nested_egor
has_ego_design(x)

strip_ego_design(x)

Arguments

x
an egor object.

...
arguments to be passed to methods

value
a list of arguments to srvyr::as_survey_design() specifying the sampling design for the egos. If the arguments are formulas, they can refer to columns (ego attributes) of x. NULL clears design information.

Note

This can be useful for adjusting or re-initializing the ego design information after the underlying ego attributes had been modified.