R Package for importing and analyzing ego-centered-network data

Tools for importing, analyzing and visualizing ego-centered network data. Supports several data formats, including the export formats of 'EgoNet', 'EgoWeb 2.0' and 'openeddi'. An interactive (shiny) app for the intuitive visualization of ego-centered networks is provided. Also included are procedures for creating and visualizing Clustered Graphs (Lerner 2008 doi:10.1109/PACIFICVIS.2008.4475458 ).

Details

When analyzing ego-centered network data it is common to either include or exclude data on ego. By default egor excludes ego for most analytic and visual functions and offers to include ego with the argument include.ego. In order for egor to automatically detect which ego-level variables correspond to the alter-variables, the variables on the different levels need to be named exactly the same. Alternatively the functions that have an include.ego argument have additional arguments that allow to specify the name of the ego-level variable in question.

For further information see package webpage or GitHub.

See also

Author

Maintainer: Till Krenz egor@tillt.net

Authors:

Pavel N. Krivitsky pavel@statnet.org

Raffaele Vacca r.vacca@ufl.edu

Michal Bojanowski michal2992@gmail.com (ORCID)

Andreas Herz herzand@uni-hildesheim.de

Other contributors:

Markus Gamper m.gamper@uni-koeln.de [contributor]

Christopher McCarty ufchris@ufl.edu [contributor]