A selective subset of GSS 2004 data

This is a selective subset of General Social Survey 2004 data containing variables from network questions. See Details for description how this particular subset was selected. The data has a near 0 research value, it is provided to illustrate the functions in egor package.

Usage

gss2004

Format

A tibble with 499 rows and the variables listed below. Data was imported from SPSS file and are labelled. Functions in the labelled package can be used to handle them.

Variables:

id

Case ID

vpsu, vstrat, wtssall

Design variables and weight

age

Ego's age in years

race

Ego's race. 1=white, 2=black, 3=other

sex

Ego's sex. 1=male, 2=female

marital

Ego's marital status. 1=married, 2=widowed, 3=divorced, 4=separated, 5=never married

numgiven

Number of alters mentioned

age[1-5]

Alter's age in years

race[1-5]

Alter's race. 1=asian, 2=black, 3=hispanic, 4=white, 5=other

sex[1-5]

Alter's sex. 1=male, 2=female

spouse[1-5]

Whether alter is a spouse of ego. 1=mentioned, 2=not mentioned

close[1-4][2-5]

How close are the two alters according to ego. 1=especially close, 2=know each other, 3=total strangers

Source

General Social Survey data at NORC: https://gss.norc.org/get-the-data.html

Details

This dataset was created from original GSS 2004 data for illustrative purposes such that (1) it is small and (2) contains just enough variation in respondent's personal networks to illustrate various functions in the package. It is essentially a stratified sample from original data (1472 cases). Strata correspond to groups of cases created from unique combinations of values on the following ego variables: age (3 categories), race, sex, marital, numgiven. At most 2 cases were sampled from each stratum via simple random sampling with replacement.