Transnational personal communities of social support of German migrants in Great Britain

This is an egor object derived from a subset of the data of a personal network study on support relationships German migrants living in the UK maintain. The data was collected in 2010 using respondent driven sampling (snowball sampling). While the number of alters the respondents were allowed to enter was not limited, only a random subsample of up to eight alters were selected for the alter name interpreter and alter-alter tie questions. This data set contains the data for 50 of the originally 234 egos.

transnat

alter_df

ego_df

Format

transnat: an egor object of 50 egos.

alter_df: alter data.frame of the transnat dataset.

ego_df: ego data.frame of the transnat dataset.

Details

The questionnaire used seven name generators:

From time to time, people rely on other people's advice and opinions to help them find their way in life better. In the last 12 months, who have you sought advice from when it came to important decisions, for example about your family or work? (emotional)

In the last 12 months who has done little jobs and favors for you or helped you, for example in filling in forms or moving home? (instrumental)

In the past year, who have you turned to when you felt down and wanted someone to talk to? (emotional)

In the last 12 months, who have you borrowed money from? (instrumental)

In the past year, who have you spent your free time with or shared a hobby? ( social companionship)

In the past year who have you had disagreements or arguments with (e.g. about everyday affairs, money or property)? (conflict)

Who has let you know that you can rely on them (e.g. that they will always be there for you if you need help)? (emotional).

References

Herz, A. (2015). Relational constitution of social support in migrants' transnational personal communities. Social Networks, 40 (1), S. 64-74.

Herz, A. (2012). Strukturen transnationaler sozialer Unterstützung. Springer Fachmedien Wiesbaden.